+++ NEWS.md
@@ -1,3 +1,5 @@
+# cpmr (development version)
+
 # cpmr 0.0.8
 
 * Added `return_edges` argument to optionally set how to return edges in the output.